Feat : postSubscriptionTrade

+++ triple7-api/src/main/java/com/sparta/triple7api/trade/dto/response/TradeResponseDto.java
@@ -5,11 +5,11 @@ import lombok.Getter;
 @Getter
 public class TradeResponseDto {
     private final String cryptoSymbol;
-    private final Long amount;
+    private final Double amount;
     private final String billType;
     private final Long price;
 
-    public TradeResponseDto(String cryptoSymbol, Long amount, String billType, Long price) {
+    public TradeResponseDto(String cryptoSymbol, Double amount, String billType, Long price) {
         this.cryptoSymbol = cryptoSymbol;
         this.amount = amount;
         this.billType = billType;
